Increase key size and stop encrypting keys
The default openssl key size varies and is in general fairely low. So we
set more reasonable sizes here.
Beside that, this removes the -des3 flag for the client and servers
keys, given that we remove the passphrase afterwards anyways.

Docker-DCO-1.1-Signed-off-by: Johannes 'fish' Ziemke <github@freigeist.org> (github: discordianfish)